Implement test case decorator for subprocess sandbox tests
With the @util._process_isolate decorator, each so-annotated test is run in
an independent process.  Some care is taken to ensure remote test failure
tracebacks are accurately transmitted back to the testrunner without extra
noise from the remote mechanism showing up or impacting test results.

You would think this would be trivial, but Python does not allow pickling
traceback objects for no particularly good reason.

The decorator is a crude hack and might be broken by future versions of
Python, but it works in 2.7 and 3.7 in my testing.