Capture coverage data across Python 2 and 3
This improves coverage because there is a small amount of code conditional
on the major version of Python (or 3.3+, anyway).  Branch coverage increases
from 68% to 72%, and line coverage from 83% to 85%.

The Cirrus CI task is also simplified nicely.  It is probably redundant to
have both a test and CI task, as we run all tests under CI.  Left for future
work.