Fix bugs identified by missing code coverage
Both the Rights and Fcntls caches were broken prior to this revision.

Ioctls initialization from fd was broken due to wrong level of whitespace (a
weakness in Python).

Finally, the compat33 os.listdir replacement would recurse until running off
the stack if used as directed.  Fix by caching a reference to the original
method in the compat33 class.